server/requirements.txt
flask
flask-cors
firebase
google-cloud-vision
google-cloud-storage
google-cloud-firestore
